requirements.txt
Appium-Python-Client==2.0.0
attrs==23.1.0       
certifi==2023.7.22  
cffi==1.16.0        
colorama==0.4.6     
cryptography==41.0.4
h11==0.14.0         
idna==3.4           
iniconfig==2.0.0    
outcome==1.2.0
packaging==23.2
pluggy==1.3.0
pycparser==2.21
pyOpenSSL==23.2.0
PySocks==1.7.1
pytest==7.4.2
pytest-ordering==0.6
selenium==4.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.2
trio-websocket==0.11.1
urllib3==1.26.17
urllib3-secure-extra==0.1.0
wsproto==1.2.0
